chore: new changes

+++ actions/doctors/delete-doctor.ts
@@ -1,6 +1,10 @@
 import { supabaseClient } from '@/lib/supabase';
 
 export async function deleteDoctor(id: string | number) {
+  if (process.env.NODE_ENV !== 'development') {
+    throw new Error('دسترسی محدود شده است!');
+  }
+
   const { error } = await supabaseClient.from('doctor').delete().eq('id', id);
 
   if (error) {
